refactor: code review cleanup, add missing tests, update docs
Minor fixes from code review:
- CupScoring/CupDefinition/MapDefinition: add final, defensive copies
- MinestomPhaseScheduler: document async limitation, use Math.toIntExact
- MinestomEventRegistrar: warn on incompatible listener types
- AnvilMapInstanceService: document commonPool usage
- PlayerServiceImpl: remove duplicate GameMode set, add comment
- MinestomLobbyPhase/EndPhase: extract shared UI logic to PhaseUiHelper
- GameHud: use Component API instead of String.format

New tests (21 added, total 70):
- GameLoopSystemTest: 8 tests for predicted position + ring detection
- GamePhaseFactoryTest: 5 tests for phase series creation
- CupScoringTest: 9 tests for cup-level score aggregation

Documentation:
- CLAUDE.md: add server module, build commands, Minestom dependency

+++ server/src/main/java/net/elytrarace/server/platform/MinestomPhaseScheduler.java
@@ -9,13 +9,17 @@ import net.minestom.server.timer.TaskSchedule;
 /**
  * Minestom implementation of {@link PhaseScheduler}.
  * Uses Minestom's scheduler to run repeating tasks at tick-based intervals.
+ * <p>
+ * Note: The {@code async} parameter in {@link #runRepeating(Runnable, long, boolean)} is
+ * ignored because Minestom's tick scheduler is single-threaded by design and does not
+ * support async task execution. All tasks run on the tick thread.
  */
 public final class MinestomPhaseScheduler implements PhaseScheduler {
 
     @Override
     public PhaseTask runRepeating(Runnable task, long intervalTicks, boolean async) {
         Task scheduledTask = MinecraftServer.getSchedulerManager()
-                .scheduleTask(task, TaskSchedule.immediate(), TaskSchedule.tick((int) intervalTicks));
+                .scheduleTask(task, TaskSchedule.immediate(), TaskSchedule.tick(Math.toIntExact(intervalTicks)));
 
         return new MinestomPhaseTask(scheduledTask);
     }
